+++ common/src/main/java/dev/obscuria/tooltips/client/tooltip/element/icon/AccentSpinIcon.java
@@ -5,11 +5,17 @@ import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import dev.obscuria.fragmentum.util.easing.Easing;
 import dev.obscuria.tooltips.client.TooltipState;
+import dev.obscuria.tooltips.client.tooltip.element.SoundTemplate;
 import dev.obscuria.tooltips.client.tooltip.element.Transform;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.util.Mth;
 
-public record AccentSpinIcon(Transform transform) implements TooltipIcon {
+import java.util.Optional;
+
+public record AccentSpinIcon(
+        Transform transform,
+        Optional<SoundTemplate> sound
+) implements TooltipIcon {
 
     public static final Codec<AccentSpinIcon> CODEC;
 
@@ -23,6 +29,7 @@ public record AccentSpinIcon(Transform transform) implements TooltipIcon {
         pushTransform(state, transform, graphics, x, y);
         graphics.renderItem(state.stack, 0, 0);
         popTransform(graphics);
+        sound.ifPresent(state::maybePlayIconSound);
     }
 
     @Override
@@ -44,7 +51,8 @@ public record AccentSpinIcon(Transform transform) implements TooltipIcon {
 
     static {
         CODEC = RecordCodecBuilder.create(codec -> codec.group(
-                Transform.CODEC.fieldOf("transform").forGetter(AccentSpinIcon::transform)
+                Transform.CODEC.fieldOf("transform").forGetter(AccentSpinIcon::transform),
+                SoundTemplate.CODEC.optionalFieldOf("sound").forGetter(AccentSpinIcon::sound)
         ).apply(codec, AccentSpinIcon::new));
     }
 }